packages/fixer-cmip7/pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["readme", "version"]
dependencies = ["cf_xarray", "fixer", "numpy", "pint", "xarray", "pyyaml"]
description = "A package for making CMIP7 data analysis ready."
license = "Apache-2.0"
license-files = ["LICENSE", "NOTICE"]
name = "fixer-cmip7"
requires-python = ">=3.12"

[project.entry-points.'fixer.plugins']
cmip7 = "fixer_cmip7:fixer"

[project.optional-dependencies]
test = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "cftime",     # TODO: update tests to use a numpy array instead.
]

[tool.setuptools]
packages = ["fixer_cmip7"]

[tool.setuptools.dynamic]
readme = { file = "README.md", content-type = "text/markdown" }

[tool.setuptools_scm]
version_scheme = "release-branch-semver"

# Configure tests

[tool.pytest.ini_options]
addopts = ["-ra", "--strict-config", "--strict-markers"]
minversion = "6"
testpaths = ["fixer_cmip7/tests"]
xfail_strict = true

[tool.coverage.run]
parallel = true
source = ["fixer_cmip7"]

[tool.coverage.report]
exclude_lines = ["if __name__ == .__main__.:", "if TYPE_CHECKING:"]
